chore: Rename css.js to css.ts
